Crate mainnet_lib

Rust implementation of Cardano various tools

[DbSyncInstance] - mock of db sync based on json file rather than postgres db Network - mock of cardano network in memory Wallet - api implementation of cardano wallet which is capable of sending and signing registration transactions

Structs

InMemoryDbSync

Mock of real cardano db sync. At this moment we only stores transactions metadata as the only purpose of existance for this struct is to provide catalyst voting registrations Struct can be persisted and restored from json file using serde_json.

METADATUM_1

metadatum label 1 constant

METADATUM_2

metadatum label 2 constant

METADATUM_3

metadatum label 3 constant

METADATUM_4

metadatum label 4 constant

MainnetNetwork

Mainnet network mock. It holds current block date as indicator on which epoch and slot mainnet network currently is. Also struct can have multiple db sync which once register can be notified on incoming transactions.

MainnetNetworkBuilder

Cardano Network state builder, responsible to create a given state of cardano network which will be an input for snapshot

MainnetWallet

Represents Cardano mainnet wallet which generate registration transaction metadata

MainnetWalletState

Represents wallet candidate for registration. Defines wallet role (delegator/direct-voter/representative)

REGISTRATION_METADATA_LABEL

registration metadata constant

REGISTRATION_METADATA_SIGNATURE_LABEL

registration signature metadata constant

RegistrationBuilder

Responsible for building registration transaction metadata

Enums

InMemoryDbSyncError

Db sync error

Traits

GeneralTransactionMetadataInfo

Extension trait for GeneralTransactionMetadata struct

MainnetWalletStateBuilder

Wallet state builder for Network state builder is a trait which creates nice interface for defining role of particular mainnet wallet in voting round. Wallet can be a direct voter/ delegator or representative