Struct mainnet_lib::InMemoryNode

pub struct InMemoryNode { /* private fields */ }

In memory cardano node implementation. Basic design contains read write lock over ledger struct as well as background task for updating ledger

Implementations

impl InMemoryNode

pub fn settings(
&self
) -> Result<Settings, PoisonError<RwLockReadGuard<'_, Ledger>>>

Retrieves settings carved in block0

Errors

On accessing shard ledger state

pub fn push_transaction(&mut self, transaction: Transaction)

Adds new transaction to mempool

Panics

On accessing shard ledger state

pub fn push_transactions(&mut self, transaction: Vec<Transaction>)

Adds collection of new transactions to mempool

Panics

On accessing shard ledger state

pub fn wait_for_txs_to_be_in_block(&self)

Waits for all transactions from mempool to be included in block

Panics

On accessing shard ledger state

pub fn start(block0: Block0) -> Self

Starts node process

Panics

On accessing ledger state

pub fn start_from_ledger(ledger: Ledger) -> Self

Starts node process from ledger

Panics

On accessing ledger state