Trait mainnet_lib::GeneralTransactionMetadataInfo

pub trait GeneralTransactionMetadataInfo {
    fn to_json_string(
        &self,
        schema: MetadataJsonSchema
    ) -> Result<String, JsonConversionError>;
    fn from_json_string(
        json: &str,
        schema: MetadataJsonSchema
    ) -> Result<Self, JsonConversionError>
    where
        Self: Sized;
    fn delegations(&self) -> Vec<u8>;
    fn stake_public_key(&self) -> PublicKey;
    fn reward_address(&self) -> Address;
    fn signature(&self) -> Ed25519Signature;
    fn registration_blake_256_hash(&self) -> Blake2b256;
}

Extension trait for GeneralTransactionMetadata struct

Required Methods

fn to_json_string(
&self,
schema: MetadataJsonSchema
) -> Result<String, JsonConversionError>

Serialize to json string. Extension allows to groups more than one metadata entry and serialize then in batch

Errors

On incorrect schema

fn from_json_string(
json: &str,
schema: MetadataJsonSchema
) -> Result<Self, JsonConversionError>where
Self: Sized,

Deserialize from json string. Extension allows to groups more than one metadata entry and deserialize then in batch

Errors

On incorrect schema

fn delegations(&self) -> Vec<u8>

Get delegations part as bytes

fn stake_public_key(&self) -> PublicKey

Get stake public key from metadata

fn reward_address(&self) -> Address

Get reward address part from metadata

fn signature(&self) -> Ed25519Signature

Get signature from metadata

fn registration_blake_256_hash(&self) -> Blake2b256

Get registration as hash