Trait mainnet_lib::wallet_state::ExternalProvider

pub trait ExternalProvider {
    type Error;

    // Required method
    fn download_state_from_network<'life0, 'life1, 'async_trait>(
        &'life0 self,
        actors: &'life1 [Actor]
    ) -> Pin<Box<dyn Future<Output = Result<Vec<MainnetWalletState>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait for retrieving information about address registrations from network

Required Associated Types

type Error

Error which can be returned on any issue with retrieving information

Required Methods

fn download_state_from_network<'life0, 'life1, 'async_trait>( &'life0 self, actors: &'life1 [Actor] ) -> Pin<Box<dyn Future<Output = Result<Vec<MainnetWalletState>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Downloads MainnetWalletState from network