Trait mainnet_lib::wallet_state::MainnetWalletStateBuilder

pub trait MainnetWalletStateBuilder {
    fn as_representative(&self) -> MainnetWalletState;
    fn as_direct_voter(&self) -> MainnetWalletState;
    fn as_direct_voter_on_slot_no(&self, slot_no: u64) -> MainnetWalletState;
    fn as_delegator(
        &self,
        delegations: Vec<(&CardanoWallet, u8)>
    ) -> MainnetWalletState;
    fn as_delegator_on_slot_no(
        &self,
        delegations: Vec<(&CardanoWallet, u8)>,
        slot_no: u64
    ) -> MainnetWalletState;
}

Wallet state builder for Network state builder is a trait which creates nice interface for defining role of particular mainnet wallet in voting round. Wallet can be a direct voter/ delegator or representative

Required Methods

fn as_representative(&self) -> MainnetWalletState

wallet registered as representative. This is simplification and wallet catalyst key is added to in memory list instead of going through public representative registration process

fn as_direct_voter(&self) -> MainnetWalletState

wallet registers as direct voter, meaning it will send self-delegation registration

fn as_direct_voter_on_slot_no(&self, slot_no: u64) -> MainnetWalletState

wallet registers as direct voter, meaning it will send self-delegation registration with given nonce = slot_no

fn as_delegator(
&self,
delegations: Vec<(&CardanoWallet, u8)>
) -> MainnetWalletState

wallet registers as delegator, meaning it will send delegation registration

fn as_delegator_on_slot_no(
&self,
delegations: Vec<(&CardanoWallet, u8)>,
slot_no: u64
) -> MainnetWalletState

wallet registers as delegator, meaning it will send delegation registration with given nonce = slot_no