Struct mainnet_lib::CardanoWallet

pub struct CardanoWallet { /* private fields */ }

Represents Cardano mainnet wallet which is able to generate registration transaction metadata

Implementations

impl CardanoWallet

pub fn new(stake: u64) -> Self

Creates new wallet with given ada. Currently wallet is purely used for testing purposes, therefore we treat stake as arbitrary number not connected to any blockchain state.

Panics

Panics on key generation error

pub fn new_with_rng<T: RngCore + CryptoRng>(stake: u64, rng: T) -> Self

Creates new wallet with given ada and rng. Currently wallet is purely used for testing purposes, therefore we treat stake as arbitrary number not connected to any blockchain state.

Panics

Panics on key generation error

pub fn reward_address(&self) -> RewardAddress

Rewards address based on stake public key

pub fn stake_credential(&self) -> StakeCredential

Stake address based on stake public key

pub fn payment_credential(&self) -> StakeCredential

Payment address based on stake public key

pub fn stake_public_key(&self) -> PublicKey

Cardano stake public key

pub fn catalyst_secret_key(&self) -> SecretKey<Ed25519Extended>

Catalyst secret key

pub fn catalyst_public_key(&self) -> Identifier

Catalyst public key

pub fn address(&self) -> BaseAddress

Mainnet address

pub fn catalyst_address(&self) -> Address

Catalyst address

pub fn generate_voting_registration(
&self,
delegations: Delegations,
slot_no: u64
) -> Transaction

Creates voting registration metadata according to Cip-36 on given absolut slot number and based on Delegations object which can be either legacy or delegation.

pub fn generate_delegated_voting_registration(
&self,
delegations: Vec<(Identifier, u32)>,
slot_no: u64
) -> Transaction

Creates delegated voting registration metadata according to Cip-36 on given absolut slot number.

pub fn generate_direct_voting_registration(&self, slot_no: u64) -> Transaction

Creates direct (a.k.a self) voting registration metadata according to Cip-36 on given absolut slot number.

pub fn stake(&self) -> u64

current amount of ada

pub fn payment_key(&self) -> &PrivateKey

private payment key

pub fn stake_key(&self) -> &PrivateKey

private stake key